update: extract away AdminList to be a reusable component
